Major refactoring pt1
Breaks everything but it was worth it. This separates the Signature
class from the search logic (now implemented in "Filters" classes). It
gives us much more flexibility when creating signatures and makes the
code smaller. Filters will also be useful when we need to display
something that isn't a vuln/signature ("show me all the filesystem
calls"). It should also be as readable (hopefully). Right now HTML and
NSData signatures don't work.