Fix /usr/libexec/cydia/firmware.sh not finishing
Details:
The new firmware.sh version will retry `gssc` process until it doesn't return a string in both stderr and stdout that contains `(null)`. Since Introspy hooks into all processes, Introspy is injected into `gssc`, it doesn't find a bundle ID and thus prints `(null)` in the logs and causing firmware.sh to keep retrying forever using battery and blocking firmware.sh's completion.

Fixed this by getting the process name instead if the bundle ID is not found. I understand this is a problem in firmware.sh also and will occur with other tweaks that print `(null)` into the logs and that should be fixed, too.